Begin JS, slight HTML edit
In this commit, I've started laying the groundwork
for the functionality of this webpage using JS. I
first added query selectors to create global
variables for the gameboard (a container div), the
div container for the game results, and my two
buttons, to which I've added ids for quit (which
will reset the game, including the players'
inputted names) and restart (which will restart
the game with the same inputted player names).
Still in the groundwork stage, I may refactor this
a little bit, given some research on a couple of
ideas, before moving forward.