Version II: refactor HTML, scripts folder
In this commit, I have overhauled and refactored
my HTML for a new approach, based on research I
have gathered about different solutions to this
coding prompt. This version will enable me to,
like I had discussed in previous commits, start
the website with a form for each player to input a
name for greater interactivity while playing the
game.

For this solution to be as effective and
efficient as  possible, I am using multiple JS
files. I created these new JS files in their own
folder titled scripts. I then fleshed out each
JS file, ensuring that export and import lines are
placed so that the players and cells scripts can
be used within the app script.

Not included in this commit is a set of temporary
styling implemented in my stylesheet, which I used
in order to test the functionality of the JS I
have implemented. For the sake of not having to go
over and tediously refactor my CSS to style the
way I actually want, I have left the temporary
styles out of this commit. Will return to styling
in future commits.

Once I could be assured of the successful
functionality of my scripts code, I cleaned up the
formatting of app.js, ensuring that it is laid out
in such a way that is readable for other devs.