Font corrections
In this commit, I've removed the Decovar font, as
I could not get it to function properly by
self-hosting. After testing a couple of other
fonts, I've decided to use the Jose Carlos Latorre
font for all text within the website, leaving the
research to greater understand the causes of this
error for a later time.