Correct js const TypeError, CSS accessibility fix
In this commit, I corrected a reference to a
const in app.js which threw a TypeError forcing
the function which resets the board to fail in
later testing. I also changed a color value in the
CSS upon recognizing that the contrast of a text
element relative to the background image was not
WCAG accessible or clear.